Move all slices wihtin src/redux/slices, update importings
